ДЗ по четвертому уроку + исправлен недочет в репозитори и сервисе (был ошибочный тип данных  поля id задан)

+++ src/main/java/ru/gb/pugacheva/homework3springboot/homework3springboot/repositories/ProductRepository.java
@@ -32,7 +32,7 @@ public class ProductRepository {
         products.add(product);
     }
 
-    public Product findById (Long id){
+    public Product findById (int id){
         return products.stream().filter(p -> p.getId()==id).findFirst().get();
     }
